Force canvas/text input focus on iOS web browsers
